FEATURE: Move AjaxWidgetComponent to middleware
